An Analysis of Kickstarter Campaigns
The Project: Kickstarter Campaigns works with Excel, and how to arrange dataset to the specific needs that you are looking for in a huge dataset. Such as customizing, filtering, adding visuals, and etc. to Excel. By analyzing the data step by step on Excel, it helped generate insights to help with Louise project, which was finding how much to fund for a theater campaign.